go: isolate lib from cli, to allow importing the lib
