+++ full-flow.sh
@@ -8,12 +8,12 @@ cargo test --release test_pod_flow -- --nocapture
 if [ ! -d "tmp/groth-artifacts" ]; then
 	echo -e "\n================================================================"
 	echo "- [generating groth16 trusted setup]: go run main.go -t"
-	go run main.go -t
+	go run cli/main.go -t
 fi
 
 echo -e "\n================================================================"
 echo "- [generating groth16 proof & solidity contract]: go run main.go -p -s"
-go run main.go -p -s
+go run cli/main.go -p -s
 
 echo -e "\n================================================================"
 echo "- proof, verifying key and the Solidity smart contract to verify the proofs can be found at 'outputs' directory"
